requirements.txt
streamlit==1.38.0
pandas
plotly
sqlalchemy
google-generativeai
python-dotenv
requests
prophet